requirements.txt
-i https://pypi.org/simple
atprototools==0.0.17
certifi==2024.2.2 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
idna==3.7 ; python_version >= '3.5'
python-dotenv==1.0.1
requests==2.32.2 ; python_version >= '3.8'
urllib3==2.2.1 ; python_version >= '3.8'
